create DeliveryService interface, migrate ExampleDeliveryService
