+++ app/src/main/java/dev/itsvic/parceltracker/api/GLS.kt
@@ -81,4 +81,3 @@ internal data class GLSHistoryAddress(
 internal data class GLSProgress(
     val statusInfo: String,
 )
-
