+++ app/src/main/java/dev/itsvic/parceltracker/ui/components/ParcelRow.kt
@@ -25,7 +25,7 @@ import androidx.compose.ui.unit.sp
 import dev.itsvic.parceltracker.R
 import dev.itsvic.parceltracker.api.Service
 import dev.itsvic.parceltracker.api.Status
-import dev.itsvic.parceltracker.api.serviceToHumanString
+import dev.itsvic.parceltracker.api.getDeliveryServiceName
 import dev.itsvic.parceltracker.api.statusToHumanString
 import dev.itsvic.parceltracker.db.Parcel
 import dev.itsvic.parceltracker.ui.theme.ParcelTrackerTheme
@@ -72,7 +72,7 @@ fun ParcelRow(parcel: Parcel, status: Status?, onClick: () -> Unit) {
             Text(parcel.humanName, color = MaterialTheme.colorScheme.onBackground)
 
             Text(
-                "${stringResource(serviceToHumanString[parcel.service]!!)}: ${parcel.parcelId}",
+                "${stringResource(getDeliveryServiceName(parcel.service)!!)}: ${parcel.parcelId}",
                 fontSize = 12.sp,
                 color = MaterialTheme.colorScheme.onSurfaceVariant
             )
